+++ R/KNN.R
@@ -122,8 +122,8 @@ fitKNN <- function(trainData, modelSettings, search = 'none', analysisId ){
   result <- list(
     model = indexFolder,
     
-    preprocess = list(
-      featureEngineering = attr(trainData$covariateData, "metaData")$featureEngineering,#learned mapping
+    preprocessing = list(
+      featureEngineering = attr(trainData, "metaData")$featureEngineering,#learned mapping
       tidyCovariates = attr(trainData$covariateData, "metaData")$tidyCovariateDataSettings,  #learned mapping
       requireDenseMatrix = F
     ),