+++ tests/testthat/test-learningCurves.R
@@ -19,18 +19,18 @@ context("LearningCurves")
 # learningCurve 
 learningCurve <- PatientLevelPrediction::createLearningCurve(
   plpData = plpData,
-  outcomeId = 2, parallel = T, cores = 3,
+  outcomeId = 2, parallel = F, cores = 3,
   modelSettings = setLassoLogisticRegression(),
   saveDirectory =  file.path(saveLoc, 'lcc'),
   splitSettings = createDefaultSplitSetting(testFraction = 0.2), 
-  trainFractions = c(0.6,0.7,0.8),
+  trainFractions = c(0.6, 0.7, 0.8),
   trainEvents = NULL,
   preprocessSettings = createPreprocessSettings(
     minFraction = 0.001,
     normalize = T
-  )
+  ),
+  logSettings = createLogSettings(verbosity="TRACE", timeStamp = T, 'runPLPLog')
 )
-
 test_that("learningCurve output correct", {